+++ Sources/RedBlackTreeModule/RedBlackTreeDictionary.swift
@@ -332,23 +332,6 @@ extension RedBlackTreeDictionary {
   }
 }
 
-extension RedBlackTreeDictionary {
-
-  /// - Complexity: O(1)
-  @inlinable
-  @inline(__always)
-  public var keys: Keys {
-    .init(tree: __tree_, start: __tree_.__begin_node, end: __tree_.__end_node())
-  }
-
-  /// - Complexity: O(1)
-  @inlinable
-  @inline(__always)
-  public var values: Values {
-    .init(tree: __tree_, start: __tree_.__begin_node, end: __tree_.__end_node())
-  }
-}
-
 // MARK: - Range Accessing Keys and Values
 
 extension RedBlackTreeDictionary {
@@ -576,19 +559,6 @@ extension RedBlackTreeDictionary {
   @inline(__always)
   @discardableResult
   public mutating func remove(at index: Index) -> KeyValue {
-    _ensureUnique()
-    guard let element = ___remove(at: index.rawValue) else {
-      fatalError(.invalidIndex)
-    }
-    return element
-  }
-
-  /// - Important: 削除後は、インデックスが無効になります。
-  /// - Complexity: O(1)
-  @inlinable
-  @inline(__always)
-  @discardableResult
-  public mutating func remove(at index: RawIndex) -> KeyValue {
     _ensureUnique()
     guard let element = ___remove(at: index.rawValue) else {
       fatalError(.invalidIndex)
@@ -893,20 +863,6 @@ extension RedBlackTreeDictionary.SubSequence: Sequence, Collection, Bidirectiona
 extension RedBlackTreeDictionary {
 
   public typealias Indices = Tree.Indices
-}
-
-// MARK: - Raw Index Sequence
-
-extension RedBlackTreeDictionary {
-
-  /// RawIndexは赤黒木ノードへの軽量なポインタとなっていて、rawIndicesはRawIndexのシーケンスを返します。
-  /// 削除時のインデックス無効対策がイテレータに施してあり、削除操作に利用することができます。
-  /// - Complexity: O(1)
-  @inlinable
-  @inline(__always)
-  public var rawIndices: RawIndexSequence<Tree> {
-    RawIndexSequence(tree: __tree_, start: __tree_.__begin_node, end: __tree_.__end_node())
-  }
 }
 
 // MARK: - ExpressibleByDictionaryLiteral