+++ Sources/RedBlackTreeModule/RedBlackTreeMultiSet.swift
@@ -330,19 +330,6 @@ extension RedBlackTreeMultiSet {
   @inlinable
   @inline(__always)
   @discardableResult
-  public mutating func remove(at index: RawIndex) -> Element {
-    _ensureUnique()
-    guard let element = ___remove(at: index.rawValue) else {
-      fatalError(.invalidIndex)
-    }
-    return element
-  }
-
-  /// - Important: 削除後は、インデックスが無効になります。
-  /// - Complexity: O(1)
-  @inlinable
-  @inline(__always)
-  @discardableResult
   public mutating func remove(at index: Index) -> Element {
     _ensureUnique()
     guard let element = ___remove(at: index.rawValue) else {
@@ -656,20 +643,6 @@ extension RedBlackTreeMultiSet.SubSequence: Sequence, Collection, BidirectionalC
 extension RedBlackTreeMultiSet {
 
   public typealias Indices = Tree.Indices
-}
-
-// MARK: - Raw Index Sequence
-
-extension RedBlackTreeMultiSet {
-
-  /// RawIndexは赤黒木ノードへの軽量なポインタとなっていて、rawIndicesはRawIndexのシーケンスを返します。
-  /// 削除時のインデックス無効対策がイテレータに施してあり、削除操作に利用することができます。
-  /// - Complexity: O(1)
-  @inlinable
-  @inline(__always)
-  public var rawIndices: RawIndexSequence<Tree> {
-    RawIndexSequence(tree: __tree_, start: __tree_.__begin_node, end: __tree_.__end_node())
-  }
 }
 
 // MARK: - Protocol Conformance