+++ Sources/RedBlackTreeModule/_Tree/___Tree.swift
@@ -225,28 +225,23 @@ extension ___Tree {
   @nonobjc
   @inlinable
   internal var _header: Header {
-    @inline(__always)
-    //    get { __header_ptr.pointee }
-    _read { yield __header_ptr.pointee }
-    @inline(__always)
-    _modify { yield &__header_ptr.pointee }
+    @inline(__always) _read {
+      yield __header_ptr.pointee
+    }
+    @inline(__always) _modify {
+      yield &__header_ptr.pointee
+    }
   }
 
   @nonobjc
   @inlinable
   public subscript(_ pointer: _NodePtr) -> Element {
-    @inline(__always)
-    //    get {
-    //      assert(0 <= pointer && pointer < _header.initializedCount)
-    //      return __node_ptr[pointer].__value_
-    //    }
-    _read {
-      assert(0 <= pointer && pointer < _header.initializedCount)
+    @inline(__always) _read {
+      assert(___initialized_contains(pointer))
       yield __node_ptr[pointer].__value_
     }
-    @inline(__always)
-    _modify {
-      assert(0 <= pointer && pointer < _header.initializedCount)
+    @inline(__always) _modify {
+      assert(___initialized_contains(pointer))
       yield &__node_ptr[pointer].__value_
     }
   }
@@ -454,31 +449,43 @@ extension ___Tree {
   @inlinable
   @inline(__always)
   internal func __parent_(_ p: _NodePtr) -> _NodePtr {
-    __node_ptr[p].__parent_
+    assert(!___is_null_or_end(p))
+    assert(___initialized_contains(p))
+    return __node_ptr[p].__parent_
   }
   @nonobjc
   @inlinable
   @inline(__always)
   internal func __left_(_ p: _NodePtr) -> _NodePtr {
-    p == .end ? _header.__left_ : __node_ptr[p].__left_
+    ___is_null_or_end(p) ? _header.__left_ : __node_ptr[p].__left_
   }
   @nonobjc
   @inlinable
   @inline(__always)
   internal func __right_(_ p: _NodePtr) -> _NodePtr {
-    __node_ptr[p].__right_
+    assert(!___is_null_or_end(p))
+    assert(___initialized_contains(p))
+    return __node_ptr[p].__right_
   }
   @nonobjc
   @inlinable
   @inline(__always)
   internal func __is_black_(_ p: _NodePtr) -> Bool {
-    __node_ptr[p].__is_black_
+    assert(!___is_null_or_end(p))
+    assert(___initialized_contains(p))
+    return __node_ptr[p].__is_black_
   }
   @nonobjc
   @inlinable
   @inline(__always)
   internal func __parent_unsafe(_ p: _NodePtr) -> _NodePtr {
-    __parent_(p)
+    // 現在、beginに対する__tree_prev_iter(..)が不定動作となっている
+    // 以下のコードで修正は可能だが、バランシングへの性能影響が大きいので
+    // 呼び出し側でチェックする運用となっている
+    // ___is_null_or_end(p) ? .nullptr : __parent_(p)
+    assert(!___is_null_or_end(p))
+    assert(___initialized_contains(p))
+    return __parent_(p)
   }
 }
 
@@ -488,12 +495,16 @@ extension ___Tree {
   @inlinable
   @inline(__always)
   internal func __is_black_(_ lhs: _NodePtr, _ rhs: Bool) {
+    assert(!___is_null_or_end(lhs))
+    assert(___initialized_contains(lhs))
     __node_ptr[lhs].__is_black_ = rhs
   }
   @nonobjc
   @inlinable
   @inline(__always)
   internal func __parent_(_ lhs: _NodePtr, _ rhs: _NodePtr) {
+    assert(!___is_null_or_end(lhs))
+    assert(___initialized_contains(lhs))
     __node_ptr[lhs].__parent_ = rhs
   }
   @nonobjc
@@ -503,6 +514,7 @@ extension ___Tree {
     if lhs == .end {
       _header.__left_ = rhs
     } else {
+      assert(___initialized_contains(lhs))
       __node_ptr[lhs].__left_ = rhs
     }
   }
@@ -510,6 +522,8 @@ extension ___Tree {
   @inlinable
   @inline(__always)
   internal func __right_(_ lhs: _NodePtr, _ rhs: _NodePtr) {
+    assert(!___is_null_or_end(lhs))
+    assert(___initialized_contains(lhs))
     __node_ptr[lhs].__right_ = rhs
   }
 }
@@ -764,6 +778,25 @@ extension ___Tree: Tree_ForEach {
   @nonobjc
   @inlinable
   @inline(__always)
+  public func ___rev_for_each_(__p _end: _NodePtr, __l _start: _NodePtr, body: (_NodePtr) throws -> Void)
+    rethrows
+  {
+    var _current = _start
+    // __begin_nodeでの__tree_prev_iterは不定動作なので注意が必要
+    var _next = _start == _end ? _end : __tree_prev_iter(_start)
+    while _current != _end {
+      _current = _next
+      _next =  _next == _end ? _end : __tree_prev_iter(_next)
+      try body(_current)
+    }
+  }
+}
+
+extension ___Tree {
+  
+  @nonobjc
+  @inlinable
+  @inline(__always)
   internal func ___for_each(
     __p: _NodePtr, __l: _NodePtr, body: (_NodePtr, inout Bool) throws -> Void
   )
@@ -775,27 +808,6 @@ extension ___Tree: Tree_ForEach {
       let __c = __p
       __p = __tree_next(__p)
       try body(__c, &cont)
-    }
-  }
-
-  @nonobjc
-  @inlinable
-  @inline(__always)
-  internal func ___for_each_(_ body: (Element) throws -> Void) rethrows {
-    try ___for_each_(__p: __begin_node, __l: __end_node(), body: body)
-  }
-
-  @nonobjc
-  @inlinable
-  @inline(__always)
-  internal func ___for_each_(__p: _NodePtr, __l: _NodePtr, body: (Element) throws -> Void)
-    rethrows
-  {
-    var __p = __p
-    while __p != __l {
-      let __c = __p
-      __p = __tree_next(__p)
-      try body(self[__c])
     }
   }
 }
@@ -931,7 +943,7 @@ extension ___Tree: Sequence {
   @nonobjc
   @inlinable
   @inline(__always)
-  public __consuming func makeIterator() -> ElementIterator<Tree> {
+  public __consuming func makeIterator() -> ElementIterator {
     .init(tree: self, start: __begin_node, end: __end_node())
   }
 }
@@ -989,16 +1001,6 @@ extension ___Tree: Tree_IndicesProtocol {
   @inline(__always)
   func makeIndices(start: _NodePtr, end: _NodePtr) -> Indices {
     .init(tree: self, start: start, end: end)
-  }
-}
-
-extension ___Tree: Tree_RawIndexProtocol {
-
-  @nonobjc
-  @inlinable
-  @inline(__always)
-  public func makeRawIndex(rawValue: _NodePtr) -> RawIndex {
-    .init(rawValue)
   }
 }