+++ Tests/RedBlackTreeTests/DictionaryRemoveTests.swift
@@ -50,15 +50,6 @@ final class DictionaryRemoveTests: XCTestCase {
     XCTAssertEqual(dict.remove(at: i).value, 1)
   }
 
-  #if DEBUG
-    func testRemoveAt() throws {
-      var s = [1: 1, 2: 2, 3: 3, 4: 4] as RedBlackTreeDictionary<Int, Int>
-      let i = s.firstIndex(of: 2)!
-      s.remove(at: RawIndex(i.rawValue))
-      XCTAssertEqual(s.map { $0.key }, [1, 3, 4])
-    }
-  #endif
-
   func testRemoveFirst() throws {
     var members: RedBlackTreeDictionary<Int, Int> = [1: 2, 3: 4, 5: 6, 7: 8, 9: 10]
     XCTAssertEqual(members.removeFirst().key, 1)
@@ -134,24 +125,24 @@ final class DictionaryRemoveTests: XCTestCase {
 
   func testRemoveWith___Indices() throws {
     var members = RedBlackTreeDictionary(uniqueKeysWithValues: (0..<10).map { ($0, $0 * 10) })
-    for i in members.rawIndices {
-      members.remove(at: i)
+    for i in members.___node_positions() {
+      members.___remove(at: i)
     }
     XCTAssertEqual(members.map { $0.key }, [])
   }
 
   func testRemoveWith___Indices2() throws {
     var members = RedBlackTreeDictionary(uniqueKeysWithValues: (0..<10).map { ($0, $0 * 10) })
-    members.rawIndices.forEach { i in
-      members.remove(at: i)
+    members.___node_positions().forEach { i in
+      members.___remove(at: i)
     }
     XCTAssertEqual(members.map { $0.key }, [])
   }
 
   func testRemoveWith___Indices3() throws {
     var members = RedBlackTreeDictionary(uniqueKeysWithValues: (0..<10).map { ($0, $0 * 10) })
-    members.rawIndices.reversed().forEach { i in
-      members.remove(at: i)
+    members.___node_positions().reversed().forEach { i in
+      members.___remove(at: i)
     }
     XCTAssertEqual(members.map { $0.key }, [])
   }
@@ -190,24 +181,24 @@ final class DictionaryRemoveTests: XCTestCase {
 
   func testRemoveWithSub___Indices() throws {
     var members = RedBlackTreeDictionary(uniqueKeysWithValues: (0..<10).map { ($0, $0 * 10) })
-    for i in members[2..<8].rawIndices {
-      members.remove(at: i)
+    for i in members[2..<8].___node_positions() {
+      members.___remove(at: i)
     }
     XCTAssertEqual(members.map { $0.key }, [0, 1, 8, 9])
   }
 
   func testRemoveWithSub___Indices2() throws {
     var members = RedBlackTreeDictionary(uniqueKeysWithValues: (0..<10).map { ($0, $0 * 10) })
-    members[2..<8].rawIndices.forEach { i in
-      members.remove(at: i)
+    members[2..<8].___node_positions().forEach { i in
+      members.___remove(at: i)
     }
     XCTAssertEqual(members.map { $0.key }, [0, 1, 8, 9])
   }
 
   func testRemoveWithSub___Indices4() throws {
     var members = RedBlackTreeDictionary(uniqueKeysWithValues: (0..<10).map { ($0, $0 * 10) })
-    members[2..<8].rawIndices.reversed().forEach { i in
-      members.remove(at: i)
+    members[2..<8].___node_positions().reversed().forEach { i in
+      members.___remove(at: i)
     }
     XCTAssertEqual(members.map { $0.key }, [0, 1, 8, 9])
   }