+++ Tests/RedBlackTreeTests/DictionaryTests.swift
@@ -215,16 +215,6 @@ final class DictionaryTests: XCTestCase {
     XCTAssertEqual(map[1], [2])
   }
 
-  #if DEBUG
-    func testSubscript1() throws {
-      let map: RedBlackTreeDictionary<Int, Int> = [1: 10, 2: 10, 3: 10]
-      XCTAssertEqual(map[RawIndex(0)].key, 1)
-      XCTAssertEqual(map[RawIndex(0)].value, 10)
-      XCTAssertEqual(map[2..<3][RawIndex(1)].key, 2)
-      XCTAssertEqual(map[2..<3][RawIndex(1)].value, 10)
-    }
-  #endif
-
   func testSmoke() throws {
     let b: RedBlackTreeDictionary<Int, [Int]> = [1: [1, 2], 2: [2, 3], 3: [3, 4]]
     print(b)
@@ -243,8 +233,8 @@ final class DictionaryTests: XCTestCase {
 
   func testInitUniqueKeysWithValues() throws {
     let dict = RedBlackTreeDictionary(uniqueKeysWithValues: [(1, 10), (2, 20)])
-    XCTAssertEqual(dict.keys.sorted(), [1, 2])
-    XCTAssertEqual(dict.values.sorted(), [10, 20])
+    XCTAssertEqual(dict.keys() + [], [1, 2])
+    XCTAssertEqual(dict.values() + [], [10, 20])
     XCTAssertEqual(dict[0], nil)
     XCTAssertEqual(dict[1], 10)
     XCTAssertEqual(dict[2], 20)
@@ -253,8 +243,8 @@ final class DictionaryTests: XCTestCase {
 
   func testInitUniqueKeysWithValues2() throws {
     let dict = RedBlackTreeDictionary(uniqueKeysWithValues: AnySequence([(1, 10), (2, 20)]))
-    XCTAssertEqual(dict.keys.sorted(), [1, 2])
-    XCTAssertEqual(dict.values.sorted(), [10, 20])
+    XCTAssertEqual(dict.keys() + [], [1, 2])
+    XCTAssertEqual(dict.values() + [], [10, 20])
     XCTAssertEqual(dict[0], nil)
     XCTAssertEqual(dict[1], 10)
     XCTAssertEqual(dict[2], 20)
@@ -288,8 +278,8 @@ final class DictionaryTests: XCTestCase {
     do {
       let dict = RedBlackTreeDictionary(
         [(1, 10), (1, 11), (2, 20), (2, 22)], uniquingKeysWith: { _, b in b })
-      XCTAssertEqual(dict.keys.sorted(), [1, 2])
-      XCTAssertEqual(dict.values.sorted(), [11, 22])
+      XCTAssertEqual(dict.keys() + [], [1, 2])
+      XCTAssertEqual(dict.values() + [], [11, 22])
       XCTAssertEqual(dict[0], nil)
       XCTAssertEqual(dict[1], 11)
       XCTAssertEqual(dict[2], 22)
@@ -298,8 +288,8 @@ final class DictionaryTests: XCTestCase {
     do {
       let dict = RedBlackTreeDictionary(
         [(1, 10), (1, 11), (2, 20), (2, 22)], uniquingKeysWith: { a, _ in a })
-      XCTAssertEqual(dict.keys.sorted(), [1, 2])
-      XCTAssertEqual(dict.values.sorted(), [10, 20])
+      XCTAssertEqual(dict.keys() + [], [1, 2])
+      XCTAssertEqual(dict.values() + [], [10, 20])
       XCTAssertEqual(dict[0], nil)
       XCTAssertEqual(dict[1], 10)
       XCTAssertEqual(dict[2], 20)
@@ -312,8 +302,8 @@ final class DictionaryTests: XCTestCase {
       let dict = RedBlackTreeDictionary(
         AnySequence([(1, 10), (1, 11), (2, 20), (2, 22)]),
         uniquingKeysWith: { _, b in b })
-      XCTAssertEqual(dict.keys.sorted(), [1, 2])
-      XCTAssertEqual(dict.values.sorted(), [11, 22])
+      XCTAssertEqual(dict.keys() + [], [1, 2])
+      XCTAssertEqual(dict.values() + [], [11, 22])
       XCTAssertEqual(dict[0], nil)
       XCTAssertEqual(dict[1], 11)
       XCTAssertEqual(dict[2], 22)
@@ -323,8 +313,8 @@ final class DictionaryTests: XCTestCase {
       let dict = RedBlackTreeDictionary(
         AnySequence([(1, 10), (1, 11), (2, 20), (2, 22)]),
         uniquingKeysWith: { a, _ in a })
-      XCTAssertEqual(dict.keys.sorted(), [1, 2])
-      XCTAssertEqual(dict.values.sorted(), [10, 20])
+      XCTAssertEqual(dict.keys() + [], [1, 2])
+      XCTAssertEqual(dict.values() + [], [10, 20])
       XCTAssertEqual(dict[0], nil)
       XCTAssertEqual(dict[1], 10)
       XCTAssertEqual(dict[2], 20)
@@ -794,18 +784,8 @@ final class DictionaryTests: XCTestCase {
     XCTAssertFalse(set.isValid(index: set.endIndex)) // 仕様変更。subscriptやremoveにつかえないので
     typealias Index = RedBlackTreeDictionary<Int, String>.Index
     #if DEBUG
-      XCTAssertEqual(RawIndex.unsafe(-1).rawValue, -1)
-      XCTAssertEqual(RawIndex.unsafe(5).rawValue, 5)
       XCTAssertEqual(Index.unsafe(tree: set.__tree_, rawValue: -1).rawValue, -1)
       XCTAssertEqual(Index.unsafe(tree: set.__tree_, rawValue: 5).rawValue, 5)
-
-      XCTAssertFalse(set.isValid(index: .unsafe(.nullptr)))
-      XCTAssertTrue(set.isValid(index: .unsafe(0)))
-      XCTAssertTrue(set.isValid(index: .unsafe(1)))
-      XCTAssertTrue(set.isValid(index: .unsafe(2)))
-      XCTAssertTrue(set.isValid(index: .unsafe(3)))
-      XCTAssertTrue(set.isValid(index: .unsafe(4)))
-      XCTAssertFalse(set.isValid(index: .unsafe(5)))
 
       XCTAssertFalse(set.isValid(index: .unsafe(tree: set.__tree_, rawValue: .nullptr)))
       XCTAssertTrue(set.isValid(index: .unsafe(tree: set.__tree_, rawValue: 0)))
@@ -826,20 +806,8 @@ final class DictionaryTests: XCTestCase {
     XCTAssertTrue(set.isValid(index: set.endIndex))
     typealias Index = RedBlackTreeDictionary<Int, String>.Index
     #if DEBUG
-      XCTAssertEqual(RawIndex.unsafe(-1).rawValue, -1)
-      XCTAssertEqual(RawIndex.unsafe(5).rawValue, 5)
       XCTAssertEqual(Index.unsafe(tree: set.__tree_, rawValue: -1).rawValue, -1)
       XCTAssertEqual(Index.unsafe(tree: set.__tree_, rawValue: 5).rawValue, 5)
-
-      XCTAssertFalse(set.isValid(index: .unsafe(.nullptr)))
-      XCTAssertFalse(set.isValid(index: .unsafe(0)))
-      XCTAssertTrue(set.isValid(index: .unsafe(1)))
-      XCTAssertTrue(set.isValid(index: .unsafe(2)))
-      XCTAssertTrue(set.isValid(index: .unsafe(3)))
-      XCTAssertTrue(set.isValid(index: .unsafe(4)))
-      XCTAssertTrue(set.isValid(index: .unsafe(5)))
-      XCTAssertFalse(set.isValid(index: .unsafe(6)))
-      XCTAssertFalse(set.isValid(index: .unsafe(7)))
 
       XCTAssertFalse(set.isValid(index: .unsafe(tree: set.__tree_, rawValue: .nullptr)))
       XCTAssertFalse(set.isValid(index: .unsafe(tree: set.__tree_, rawValue: 0)))