+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.25] - 2025-6-?
+### Added
+- 各種シーケンスにシーケンス生成メソッドを追加
+### Changed
+- RawIndexを用いるforEachを非公開扱いに降格
+- rawIndicesを非公開扱いに降格
+- 辞書やmultimapのkeys及びvaluesをプロパティからメソッドに変更
+
 ## [0.1.24] - 2025-6-7
 ### Changed
 - ビルド時の警告を抑制